refactored dialog component
